08_llm_attacks_multimodal/app/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
pydantic==2.8.2
python-multipart==0.0.9
Pillow==10.4.0
